code/SCP/requirements.txt
argparse; python_version < '3.2'
# jupyterlab>=3  # Optional, for running notebooks.
matplotlib
numpy>=1.21.2
pandas
pytorch-ignite~=0.4.2
scipy>=1.4.1
torch==1.5.0
typing; python_version < '3.5'
